Changes
1. Add old jquery.terminal-1.6.3.min.js from here: https://github.com/jcubic/jquery.terminal/archive/1.6.3.zip to compare.
Old jquery.terminal.min.js, this file, and this: https://insect.sh/jquery.terminal.min.js are identical.
2. Add jquery.terminal-1.9.0.min.js from https://github.com/jcubic/jquery.terminal/archive/1.9.0.zip to compare
Current jquery.terminal.min.js and this are identical.
3. Add uncompressed version jquery.terminal-1.8.0.js with version in name
from here: https://github.com/jcubic/jquery.terminal/archive/1.8.0.zip

Why? I see an error in version 1.9.0.
There is symbol '>' in the string '->' is encoded as &amp;#62; and not as HTML entities &#62;
So ampersand is encoded two times as HTML entitie.
And when string '&amp;#62;' go to span as HTML code, this is displayed as &#62; not as '>'

3. Update index.html with included this uncompressed JQuery Terminal v1.8.0
(minified v1.6.3, v1.9.0 are included but commented).

Now, all symbols displayed good, but after copy and paste I can see double symbol.