Support running Insect with Node 10
Fixes #306.

I was unable to get Node 10 to run ESM files, so I made `index.cjs` a
CommonJS file to work around the problem. I also had to downgrade
`clipboardy` and `xdg-basedir` to versions that support CommonJS
imports. Additional, the benchmark script was modified to benchmark
`index.cjs` instead of `index.js`, as that is what would actually be run
by those who install Insect.

As a result of consolidating what was previously `index.js` and
`insect.js` into one bundled and minified file `index.cjs`, the
startup time was also slightly improved. Using the modified benchmark
script with this commit, here are the results on my machine (with Node
18.7.0):

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `Startup time (insect '1+1')` | 161.4 ± 1.8 | 157.7 | 164.5 | 1.00 |

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `Startup time, stdin mode (insect < computations-1.ins)` | 180.5 ± 7.5 | 176.3 | 206.1 | 1.00 |

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `Evaluation speed (insect < computations-160.ins)` | 1047.8 ± 15.0 | 1028.2 | 1072.1 | 1.00 |

On current `master`:

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `Startup time (insect '1+1')` | 179.9 ± 1.3 | 178.2 | 183.4 | 1.00 |

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `Startup time, stdin mode (insect < computations-1.ins)` | 199.9 ± 2.0 | 196.9 | 202.5 | 1.00 |

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `Evaluation speed (insect < computations-160.ins)` | 1079.1 ± 25.9 | 1044.4 | 1124.0 | 1.00 |